Suggested changes made.
