+++ config.yaml
@@ -34,6 +34,8 @@ server:
     ## the maximum duration before timing out writes of the response.
     ## Value can be set in seconds (e.g. "5s") or milliseconds (e.g. "5000ms")
     writeTimeout: 10s
+    ## path for the health status api
+    healthCheckPath: "/health"
     ## the location of the TLS key file
 #    keyFile: <key-file>
     ## the location of the TLS certificate file
@@ -51,8 +53,6 @@ api:
     enableNotifications: false
     ## set to true to be able to override experiment bucketing. (recommended false in production)
     enableOverrides: true
-    ## endpoint for the health status api
-    healthEndPoint: "health"
     ## for details check this out.
     ## https://github.com/go-chi/cors
     # cors:
