+++ config/config_test.go
@@ -32,6 +32,7 @@ func TestDefaultConfig(t *testing.T) {
 
 	assert.Equal(t, 5*time.Second, conf.Server.ReadTimeout)
 	assert.Equal(t, 10*time.Second, conf.Server.WriteTimeout)
+	assert.Equal(t, "/health", conf.Server.HealthCheckPath)
 	assert.Equal(t, "", conf.Server.KeyFile)
 	assert.Equal(t, "", conf.Server.CertFile)
 	assert.Equal(t, []string{}, conf.Server.DisabledCiphers)
@@ -48,7 +49,6 @@ func TestDefaultConfig(t *testing.T) {
 
 	assert.Equal(t, 0, conf.API.MaxConns)
 	assert.Equal(t, "8080", conf.API.Port)
-	assert.Equal(t, "health", conf.API.HealthEndPoint)
 	assert.Equal(t, make([]OAuthClientCredentials, 0), conf.API.Auth.Clients)
 	assert.Equal(t, make([]string, 0), conf.API.Auth.HMACSecrets)
 	assert.Equal(t, time.Duration(0), conf.API.Auth.TTL)
