+++ pkg/routers/api.go
@@ -18,7 +18,6 @@
 package routers
 
 import (
-	"fmt"
 	"net/http"
 
 	"github.com/rakyll/statik/fs"
@@ -40,12 +39,10 @@ import (
 // APIOptions defines the configuration parameters for Router.
 type APIOptions struct {
 	maxConns        int
-	healthEndPoint  string
 	sdkMiddleware   func(next http.Handler) http.Handler
 	metricsRegistry *metrics.Registry
 	configHandler   http.HandlerFunc
 	activateHandler http.HandlerFunc
-	healthHandler   http.HandlerFunc
 	trackHandler    http.HandlerFunc
 	overrideHandler http.HandlerFunc
 	nStreamHandler  http.HandlerFunc
@@ -90,11 +87,9 @@ func NewDefaultAPIRouter(optlyCache optimizely.Cache, conf config.APIConfig, met
 
 	spec := &APIOptions{
 		maxConns:        conf.MaxConns,
-		healthEndPoint:  conf.HealthEndPoint,
 		metricsRegistry: metricsRegistry,
 		configHandler:   handlers.OptimizelyConfig,
 		activateHandler: handlers.Activate,
-		healthHandler:   handlers.Health,
 		overrideHandler: overrideHandler,
 		trackHandler:    handlers.TrackEvent,
 		sdkMiddleware:   mw.ClientCtx,
@@ -122,8 +117,6 @@ func WithAPIRouter(opt *APIOptions, r chi.Router) {
 	overrideTimer := middleware.Metricize("override", opt.metricsRegistry)
 	trackTimer := middleware.Metricize("track-event", opt.metricsRegistry)
 	createAccesstokenTimer := middleware.Metricize("create-api-access-token", opt.metricsRegistry)
-	healthTimer := middleware.Metricize(opt.healthEndPoint, opt.metricsRegistry)
-	healthEndPoint := fmt.Sprintf("/%s", opt.healthEndPoint)
 
 	if opt.maxConns > 0 {
 		// Note this is NOT a rate limiter, but a concurrency threshold
@@ -142,7 +135,6 @@ func WithAPIRouter(opt *APIOptions, r chi.Router) {
 		r.With(opt.oAuthMiddleware).Get("/notifications/event-stream", opt.nStreamHandler)
 	})
 
-	r.With(healthTimer).Get(healthEndPoint, opt.healthHandler)
 	r.With(createAccesstokenTimer).Post("/oauth/token", opt.oAuthHandler)
 
 	statikFS, err := fs.New()
